// apps/web/lib/fanout/abis/contracts.generated.ts
export const deployedAddresses = {
  "ausd": "0x34C2CFdE74D0edbABF2F4382CEbD9F600048d17E",
  "treasury": "0xF6117E2bbf56bb942DB279B6bec461761ee0De73",
  "batchPayout": "0xd75aCB8BffEc8F8BF38D067e0F26E263954EA783",
  "claimEscrow": "0xbD1519D8f5602cE5CFed89E505AD81717567fd78"
} as const;